requirements.txt
geopandas
matplotlib
folium
mapclassify
streamlit_folium
pydeck
pyecharts
streamlit_echarts
branca
datetime